+++ app/src/main/java/woowacourse/movie/domain/model/cinema/screen/ScreenSize.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model.cinema
+package woowacourse.movie.domain.model.cinema.screen
 
 data class ScreenSize(
     val rowSize: Int,
