+++ app/src/main/java/woowacourse/movie/presentation/model/ScreenUiModel.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.presentation.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
-import woowacourse.movie.domain.model.cinema.Screen
+import woowacourse.movie.domain.model.cinema.screen.Screen
 
 @Parcelize
 class ScreenUiModel(
