+++ app/src/test/java/woowacourse/movie/domain/model/movie/MovieTest.kt
@@ -1,12 +1,8 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.movie
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
-import woowacourse.movie.domain.model.movie.Movie
-import woowacourse.movie.domain.model.movie.Poster
-import woowacourse.movie.domain.model.movie.RunningTime
-import woowacourse.movie.domain.model.movie.ScreeningPeriod
 import java.time.LocalDate
 
 class MovieTest {
