+++ app/src/test/java/woowacourse/movie/domain/model/movie/ScreeningPeriodTest.kt
@@ -1,11 +1,10 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.movie
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
 import org.junit.jupiter.api.assertThrows
-import woowacourse.movie.domain.model.movie.ScreeningPeriod
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
